+++ db/schema.rb
@@ -55,7 +55,7 @@ ActiveRecord::Schema.define(version: 2021_05_18_133024) do
   create_table "comments", force: :cascade do |t|
     t.text "description", null: false
     t.integer "sentiment", default: 0
-    t.integer "sentiment_score", default: 0
+    t.float "sentiment_score", default: 0.0
     t.bigint "ticket_id", null: false
     t.bigint "commenter_id"
     t.datetime "created_at", precision: 6, null: false
@@ -71,7 +71,7 @@ ActiveRecord::Schema.define(version: 2021_05_18_133024) do
     t.string "name", null: false
     t.integer "status", default: 0
     t.integer "sentiment", default: 0
-    t.integer "sentiment_score", default: 0
+    t.float "sentiment_score", default: 0.0
     t.datetime "last_activity"
     t.date "due_date"
     t.bigint "creator_id"
